Add command permission validation
