+++ src/commands/credentials.js
@@ -56,23 +56,7 @@ module.exports = {
                 .setName('is_set')
                 .setDescription('Is the FCM Credentials already set for this Discord Server?')),
     async execute(client, interaction) {
-        let instance = client.readInstanceFile(interaction.guildId);
-
-        if (instance.role !== null) {
-            if (!interaction.member.permissions.has('ADMINISTRATOR') &&
-                !interaction.member.roles.cache.has(instance.role)) {
-                let role = DiscordTools.getRole(interaction.guildId, instance.role);
-                let str = `You are not part of the '${role.name}' role, therefore you can't run bot commands.`;
-                await client.interactionReply(interaction, {
-                    embeds: [new MessageEmbed()
-                        .setColor('#ff0040')
-                        .setDescription(`\`\`\`diff\n- ${str}\n\`\`\``)],
-                    ephemeral: true
-                });
-                client.log('WARNING', str);
-                return;
-            }
-        }
+        if (!await client.validatePermissions(interaction)) return;
 
         await interaction.deferReply({ ephemeral: true });
